Farm out code to separate components and remove obsolete package.
