+++ package.json
@@ -26,7 +26,6 @@
     "@types/node": "18.15.11",
     "@types/react": "18.0.31",
     "@types/react-dom": "18.0.11",
-    "date-fns": "^2.29.3",
     "dotenv": "^16.0.3",
     "lodash": "^4.17.21",
     "next": "13.2.4",
