+++ src/components/messageDashboard/index.tsx
@@ -1,15 +1,14 @@
-import React, { useCallback, useState } from "react";
-import { Box, TextField } from "@mui/material";
+import React from "react";
+import { Backdrop, Box, CircularProgress, Typography } from "@mui/material";
 
-import { LoadingButton } from "@mui/lab";
-
+import useUsers from "@customHooks/useUser";
 import useMessages from "@customHooks/useMessages";
 import useSocket from "@customHooks/useSocket";
 import { User } from "@customTypes/users";
-import useUsers from "@customHooks/useUser";
 
-import UserMessage from "./UserMessage";
 import UserOptions from "./UserOptions";
+import Messenger from "./Messenger";
+import Messages from "./Messages";
 
 type IMessageDashboard = {
   user: User;
@@ -17,79 +16,27 @@ type IMessageDashboard = {
 
 export default function MessageDashboard(props: IMessageDashboard) {
   const { user } = props;
-
   const { socket } = useSocket();
-  const {
-    messages,
-    mutate: { mutate: addMessage, isLoading: isAddingMessage },
-  } = useMessages(socket);
   const { data: users } = useUsers(socket);
-  const [messageText, setMessageText] = useState<string>("");
+  const { messages } = useMessages(socket);
 
-  const handleSubmit = useCallback(
-    async (evt: React.FormEvent<HTMLButtonElement>) => {
-      evt.preventDefault();
-      await addMessage({
-        userId: user.userId,
-        text: messageText,
-      });
-      setMessageText("");
-    },
-    [messageText, user, addMessage]
-  );
+  if (!socket) {
+    return (
+      <Backdrop
+        open
+        sx={{ color: "white", display: "flex", flexDirection: "column" }}
+      >
+        <CircularProgress color="inherit" />
+        <Typography>Connecting to the server. Please wait...</Typography>
+      </Backdrop>
+    );
+  }
 
   return (
-    <Box marginBottom="75px">
+    <Box sx={{ marginBottom: "75px" }}>
       <UserOptions user={user} />
-      {users && (
-        <Box sx={{ overflowX: "scroll", marginTop: "8px" }}>
-          {messages.map((message) => (
-            <UserMessage
-              message={message}
-              user={users.find((u) => u.userId === message.userId)!}
-              key={message.messageId}
-            />
-          ))}
-        </Box>
-      )}
-      <Box
-        sx={{
-          height: "40px",
-          display: "flex",
-          position: "fixed",
-          bottom: 0,
-          left: 0,
-          right: 0,
-          padding: "16px",
-          flexDirection: "row",
-          alignItems: "center",
-          backgroundColor: "#999",
-        }}
-      >
-        <TextField
-          placeholder="Enter a message..."
-          multiline
-          fullWidth
-          color="primary"
-          disabled={isAddingMessage}
-          sx={{
-            margin: 0,
-            backgroundColor: "white",
-          }}
-          value={messageText}
-          onChange={(evt) => setMessageText(evt.target.value)}
-        />
-        <LoadingButton
-          color="success"
-          variant="contained"
-          sx={{ marginLeft: "8px" }}
-          loading={isAddingMessage}
-          onClick={handleSubmit}
-          disabled={messageText.length === 0 || isAddingMessage}
-        >
-          Submit
-        </LoadingButton>
-      </Box>
+      {users ? <Messages users={users} messages={messages} /> : null}
+      <Messenger user={user} />
     </Box>
   );
 }
